requirements.txt
apache-airflow[fab]>=3.1,<4
psycopg2-binary>=2.9.0
asyncpg>=0.30.0
PyYAML>=6.0
